pehitungan jarak

+++ backend-api/database/migrations/2021_10_01_181134_create_workshops_table.php
@@ -27,8 +27,8 @@ class CreateWorkshopsTable extends Migration
             $table->string('city');
             $table->string('district');
             $table->string('province');
-            $table->float( 'latitude',10, 6);
-            $table->float( 'longitude',10, 6);
+            $table->decimal( 'latitude',10, 6);
+            $table->decimal( 'longitude',10, 6);
             $table->string('statusBuka')->nullable();
             $table->string('status24jam')->nullable();
             $table->boolean("isActive")->default(1);
